requirements.txt
torch
torchvision
pandas
Pillow
ipykernel
jupyter
notebook
pytorch-lightning
torchmetrics